+++ experimental/packages/api-events/src/api/events.ts
@@ -20,10 +20,10 @@ import {
   _global,
   makeGetter,
 } from '../internal/global-utils';
-import { EventEmitterProvider } from '../types/EventEmitterProvider';
-import { NOOP_EVENT_EMITTER_PROVIDER } from '../NoopEventEmitterProvider';
-import { EventEmitter } from '../types/EventEmitter';
-import { EventEmitterOptions } from '../types/EventEmitterOptions';
+import { EventLoggerProvider } from '../types/EventLoggerProvider';
+import { NOOP_EVENT_LOGGER_PROVIDER } from '../NoopEventLoggerProvider';
+import { EventLogger } from '../types/EventLogger';
+import { EventLoggerOptions } from '../types/EventLoggerOptions';
 
 export class EventsAPI {
   private static _instance?: EventsAPI;
@@ -38,54 +38,48 @@ export class EventsAPI {
     return this._instance;
   }
 
-  public setGlobalEventEmitterProvider(
-    provider: EventEmitterProvider
-  ): EventEmitterProvider {
+  public setGlobalEventLoggerProvider(
+    provider: EventLoggerProvider
+  ): EventLoggerProvider {
     if (_global[GLOBAL_EVENTS_API_KEY]) {
-      return this.getEventEmitterProvider();
+      return this.getEventLoggerProvider();
     }
 
-    _global[GLOBAL_EVENTS_API_KEY] = makeGetter<EventEmitterProvider>(
+    _global[GLOBAL_EVENTS_API_KEY] = makeGetter<EventLoggerProvider>(
       API_BACKWARDS_COMPATIBILITY_VERSION,
       provider,
-      NOOP_EVENT_EMITTER_PROVIDER
+      NOOP_EVENT_LOGGER_PROVIDER
     );
 
     return provider;
   }
 
   /**
-   * Returns the global event emitter provider.
+   * Returns the global event logger provider.
    *
-   * @returns EventEmitterProvider
+   * @returns EventLoggerProvider
    */
-  public getEventEmitterProvider(): EventEmitterProvider {
+  public getEventLoggerProvider(): EventLoggerProvider {
     return (
       _global[GLOBAL_EVENTS_API_KEY]?.(API_BACKWARDS_COMPATIBILITY_VERSION) ??
-      NOOP_EVENT_EMITTER_PROVIDER
+      NOOP_EVENT_LOGGER_PROVIDER
     );
   }
 
   /**
-   * Returns a event emitter from the global event emitter provider.
+   * Returns a event logger from the global event logger provider.
    *
-   * @returns EventEmitter
+   * @returns EventLogger
    */
-  public getEventEmitter(
+  public getEventLogger(
     name: string,
-    domain: string,
     version?: string,
-    options?: EventEmitterOptions
-  ): EventEmitter {
-    return this.getEventEmitterProvider().getEventEmitter(
-      name,
-      domain,
-      version,
-      options
-    );
+    options?: EventLoggerOptions
+  ): EventLogger {
+    return this.getEventLoggerProvider().getEventLogger(name, version, options);
   }
 
-  /** Remove the global event emitter provider */
+  /** Remove the global event logger provider */
   public disable(): void {
     delete _global[GLOBAL_EVENTS_API_KEY];
   }